address React warning, move files
