+++ frontend.env.default
@@ -9,7 +9,7 @@ REACT_APP_FHIR_RESOURCES=Questionnaire,QuestionnaireResponse,CarePlan
 ################# Launch context scope ###############################
 #for launching from other launchers e.g. smarthealth it launcher
 #REACT_APP_AUTH_SCOPES=profile roles email patient/*.read openid fhirUser patient/QuestionnaireResponse.write
-#for launching from DEM environment
+#for launching from DCW environment
 REACT_APP_AUTH_SCOPES=profile roles email patient/*.read openid
 #####################################################################
