+++ src/context/FhirClientProvider.js
@@ -3,8 +3,8 @@ import FHIR from "fhirclient";
 import Stack from "@mui/material/Stack";
 import CircularProgress from "@mui/material/CircularProgress";
 import { FhirClientContext } from "./FhirClientContext";
-import { queryPatientIdKey } from "./util/util";
-import ErrorComponent from "./components/ErrorComponent";
+import { queryPatientIdKey } from "../util/util";
+import ErrorComponent from "../components/ErrorComponent";
 
 export default function FhirClientProvider(props) {
   const [client, setClient] = useState(null);
