+++ src/launch.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import Launch from './components/LaunchComponent';
+import LaunchComponent from "./components/LaunchComponent";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Launch />
+    <LaunchComponent />
   </React.StrictMode>
 );
